+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Medium Impact Changes**
 - **Other Features**
   - [spiral/queue] Added the ability to pass headers in the `headers` parameter in the job handlers.
+  - [spiral/events] Added event interceptors.
 
 ## v3.1.0 - 2022-09-29
 - **Other Features**